pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "marp-generator"
version = "2.0.0"
description = "Automated PowerPoint generator using GPT and Marp CLI"
readme = "md/README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "requests>=2.28",
    "python-dotenv>=1.0",
]

[project.scripts]
marp-gen = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["marp_core*"]

[tool.setuptools.package-data]
"marp_core" = ["templates/*.md"]
